api/requirements.txt
PyYAML
pymongo
flask
Flask-RESTful
